wx: Builds cleanly on 64bit, thanks to mwerle
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@4886 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/ui/wxWidgets/addeditpropsheet.cpp
@@ -1130,7 +1130,7 @@ static short GetSelectedDCA(wxComboBox *pcbox, short defval)
   if (sel == wxNOT_FOUND) { // no selection - is this possible with our combobox?
     return -1;
   } else {
-    retval = reinterpret_cast<int>(pcbox->GetClientData(sel));
+    retval = reinterpret_cast<intptr_t>(pcbox->GetClientData(sel));
   }
   if (retval == defval)
     retval = -1;
